backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
python-dotenv==1.0.1
pywebpush==1.14.0
apscheduler==3.10.4
aiofiles==23.2.1
requests==2.31.0